First commit to include Telegraf
This is a work-in-progress commit replacing the previous system metric collection code with Telegraf. Packaging and releasing should work exactly the same as before. This does not include any ability to configure Telegraf from the ISD config file. It automatically submits all metrics to the Telegraf test project. Also, it could use some drying up.