Updated telegraph config to use appropriate values from the instrumentald command line and config file.
It also removes the --collector command line option because it would take too much time at the moment to make this apply to telegraph.

More specifically this makes the options `report_interval`, `debug`, and `hostname` apply to telegraph.

This includes some refactoring to make it easy to get the appropriate option regardless of where it is configured (command line, config file, or a default option, in order of precedence).