Added rake package:release task to push all packages to Package Cloud.
This includes corresponding updates to BUILD.md along with some reorganization. It also changes the rake task name for individual packages from 'push' to 'release' to match package:release.

This also fixes what seemed to be an existing issue where before a package was pushed to Package Cloud we would yank a package of the same name to ensure there wasn't an existing package that would block the release. Apparently Package Cloud automatically changes the package name in some cases depending on the distro. This means the yank fails because the package in Package Cloud has a different name than the name we used to push it. In this PR you'll see some code to transform the yanking package name to match Package Cloud's name. It seems like there's probably a better way to do this, but not being super familiar with this whole packaging system I'm not sure what that way is. I'm willing to ride with this until we learn more about it and run into this again.