Smallest change to get Ubuntu 12.04 testing working with Chef
When ISD is initially installed it attempts to start and fails due to lack of API key (which is totally reasonable). However, this causes Chef to fail before running the tests (which is not helpful).

This small change causes the template config file to be installed before attempting to install the package FOR JUST THIS ONE PLATFORM FAMILY, thus making sure that ISD will actually start and then the tests will run.

Also includes a update to the syntax of variable lookup in Chef templates, so that the api key is actually put into the template.