Smooth out test running, particularly puppet
The TEST.md directions were out of date, referencing VMWare instead of
VirtualBox. Also the puppet tests didn't work all the time,
particularly if you were testing a gem that hadn't been released to
packagecloud yet. They gave an error about hiera not working.

This updates the TEST.md directions and adds a `script/test` executable
that automatically runs everything necessary, including packaging and
running both chef and puppet tests.

This also removes hiera usage from puppet and replaces it with vagrant
directory syncing (which is what we already used for chef).

This adds the ability to test unpublished packages with puppet, and
makes that the default (like chef).