Allow MongoDB Atlas URIs, including ssl options
Previously Atlas URIs didn't work because telegraf couldn't parse the
ssl options. Here we add parsing that transforms Atlas URIs into
configuration blocks that telegraf understands.

This uses the Mongo ruby driver to parse URIs, but unfortunately
requires some shimming which you'll see in the code change.

This also updates the tests to include MongoDB configuration so we can
verify we process these into the correct telegraf config. In order to
test the telegraf config we had to move away from storing it in a
tempfile. The tempfile was getting deleted when a temporary forked
process exited, which called the tempfile's finalize method.

The general approach here is to break apart a MongoDB URI into its
multiple server components and specify a telegraf config block for
each, along with the appropriate ssl config for each.

This also adds another test to ensure instrumentald is running because
there are cases where `should be_running` passes incorrectly (notably
whenever the instrumentald init.d script is used because it always
exits with a success status, whether or not the process is running).