Install systemd script with deb and fix related issues
Previously in the deb the init.d script was included, but not the
systemd script. FPM has an option for it, so this PR adds that option.
This results in the systemd script being installed when you install the
deb. That makes interfacing through puppet easier using Service.

This uncovered an issue with the previous systemd script where the log
file option wasn't taking effect because the "foreground" command
logged to stdout.

This PR also updates the systemd script to work around that issue.