Add Ubuntu 18 and Debian 9 to test and release
Ubuntu 18 and Debian 9 are live and users are going to be asking for support for these.

Detailed discussion: https://3.basecamp.com/3093825/buckets/2065368/documents/1078567569#__recording_1106429121

This simply adds an ubuntu 18 and debian 9 image to the kitchen test file and the release pusher.  Becuase we use one rpm for all rpm-type release and one deb for all deb-type releases, we do not need to actually build any new packages, just push them under a new name.